src/com/shroggle/presentation/image/RemoveIconService.java
/*********************************************************************
 *                                                                   *
 * Copyright (c) 2007-2011 by Web-Deva.                              *
 * All rights reserved.                                              *
 *                                                                   *
 * This computer program is protected by copyright law and           *
 * international treaties. Unauthorized reproduction or distribution *
 * of this program, or any portion of it, may result in severe civil *
 * and criminal penalties, and will be prosecuted to the maximum     *
 * extent possible under the law.                                    *
 *                                                                   *
 *********************************************************************/
package com.shroggle.presentation.image;

import com.shroggle.logic.user.UsersManager;
import com.shroggle.presentation.AbstractService;
import com.shroggle.util.ServiceLocator;
import org.directwebremoting.annotations.RemoteProxy;

/**
 * @author Balakirev Anatoliy
 */
@RemoteProxy
public class RemoveIconService extends AbstractService {

    public void remove(final int iconId) {
        new UsersManager().getLogined();

        ServiceLocator.getPersistanceTransaction().execute(new Runnable() {

            @Override
            public void run() {
                ServiceLocator.getPersistance().removeIcon(iconId);
            }

        });
    }

}
